Created a new header with my basic array functions, unnecessary functions were removed from the rest of the files
