+++ FourthHomework28.09.21/ModulesAndFiles/ModulesAndFiles.c
@@ -3,30 +3,8 @@
 #include <stdlib.h>
 #include <time.h>
 #include "myQSort.h"
-
-void swap(int* firstVariable, int* secondVariable)
-{
-    *firstVariable ^= *secondVariable;
-    *secondVariable ^= *firstVariable;
-    *firstVariable ^= *secondVariable;
-}
-
-void fillArrayWithRandomElements(int theArray[], int lengthOfTheArray, int mod)
-{
-    for (int i = 0; i < lengthOfTheArray; ++i)
-    {
-        theArray[i] = rand() % mod;
-    }
-}
-
-void printArray(const int theArray[], const int lengthOfTheArray)
-{
-    for (int i = 0; i < lengthOfTheArray; ++i)
-    {
-        printf("%d ", theArray[i]);
-    }
-    printf("\n\n");
-}
+#include "testMyQSort.h"
+#include "WorkWithArrays.h"
 
 int findTrend(int theArray[], int lengthOfTheArray)
 {
@@ -52,38 +30,8 @@ int findTrend(int theArray[], int lengthOfTheArray)
     return trend;
 }
 
-bool checkThatArrayIsSorted(const int theArray[], const int lengthOfTheArray)
-{
-    for (int i = 0; i < lengthOfTheArray - 1; ++i)
-    {
-        if (theArray[i] > theArray[i + 1])
-        {
-            return false;
-        }
-    }
-    return true;
-}
-
-bool checkThatAnArraysAreIdentical(const int firstArray[],
-    const int secondArray[], const int lengthOfArrays)
-{
-    bool identical = true;
-
-    for (int i = 0; i < lengthOfArrays; ++i)
-    {
-        if (firstArray[i] != secondArray[i])
-        {
-            identical = false;
-            break;
-        }
-    }
-
-    return identical;
-}
-
-int* readArrayFromFile(char nameOfFile[], int* lengthOfTheArray, int* errorNumber)
-{
-    FILE* file = fopen(nameOfFile, "r");
+int* readArrayFromFile(char nameOfFile[], int* lengthOfTheArray, int* errorNumber) {
+    FILE *file = fopen(nameOfFile, "r");
     if (file == NULL)
     {
         *errorNumber = -1;
@@ -103,6 +51,7 @@ int* readArrayFromFile(char nameOfFile[], int* lengthOfTheArray, int* errorNumbe
     
     if (theArray == NULL)
     {
+        fclose(file1);
         *errorNumber = -2;
         return NULL;
     }
@@ -138,70 +87,6 @@ bool testReadFromFile()
     return lengthOfTheArray == 10 && checkThatAnArraysAreIdentical(arrayAnswer, arrayReading, 10);
 }
 
-bool testInsertionSort()
-{
-    int firstArray[5] = { 2, 3, 1, 5, 2 };
-    const int sortedTheFirstArray[5] = { 1, 2, 2, 3, 5 };
-    insertionSort(firstArray, 5);
-
-    int secondArray[3] = { -1, -1, -1 };
-    const int sortedTheSecondtArray[3] = { -1, -1, -1 };
-    insertionSort(secondArray, 3);
-
-    int thirdArray[10] = { 2, 5, 5, 1, 8, 3, 9, 4, 0, 7 };
-    const int sortedTheThirdArray[10] = { 0, 1, 2, 3, 4, 5, 5, 7, 8, 9 };
-    insertionSort(thirdArray, 10);
-
-    return checkThatAnArraysAreIdentical(firstArray, sortedTheFirstArray, 5) &&
-        checkThatAnArraysAreIdentical(secondArray, sortedTheSecondtArray, 3) &&
-        checkThatAnArraysAreIdentical(thirdArray, sortedTheThirdArray, 10);
-}
-
-bool testMyQSortOnRandomArrays(const int maxLengthOfTheArrays, const int numberOfArrays)
-{
-    srand((unsigned)time(NULL));
-
-    int* randomArray = (int*)calloc(maxLengthOfTheArrays, sizeof(int));
-    bool sortIsWorking = true;
-
-    if (randomArray == NULL)
-    {
-        return true;
-    }
-
-    for (int i = 0; i < numberOfArrays; ++i)
-    {
-        int lengthOfTheArray = rand() % maxLengthOfTheArrays;
-        fillArrayWithRandomElements(randomArray, lengthOfTheArray, 1000);
-        int supportElement = randomArray[0];
-        myQSort(randomArray, lengthOfTheArray);
-
-        if (!checkThatArrayIsSorted(randomArray, lengthOfTheArray))
-        {
-            printf("Wrong myQSort on the array:\n");
-            printArray(randomArray, lengthOfTheArray);
-            sortIsWorking = false;
-            break;
-        }
-    }
-    free(randomArray);
-    return sortIsWorking;
-}
-
-bool testMyQSort()
-{
-    int array1[5] = { 2, 5, 1, 6, -2 };
-    int array2[10] = { 3, 1, 1, 7, -2, 6, -2, -6, 20, 10 };
-    int array3[20] = { 5, 2, 5, 2, 6, 2, 6, 0, -3, 12,
-        22, -22, 3, 1, 5, -6, 0, 222, 1, -50 };
-    myQSort(array1, 5);
-    myQSort(array2, 10);
-    myQSort(array3, 20);
-
-    return checkThatArrayIsSorted(array1, 5) && checkThatArrayIsSorted(array2, 10) &&
-        checkThatArrayIsSorted(array3, 20);
-}
-
 bool testFindTrend()
 {
     int array1[10] = { 0, 1, 2, 3, 5, 5, 5, 6, 9, 9 };
@@ -213,8 +98,7 @@ bool testFindTrend()
 int main()
 {
     srand((unsigned)time(NULL));
-    if (!testMyQSort() || !testInsertionSort() || !testMyQSortOnRandomArrays(10, 10) ||
-        !testFindTrend() || !testReadFromFile())
+    if (!testMyQSort() || !testFindTrend() || !testReadFromFile())
     {
         printf("Tests failed(\n");
         return -1;
