+++ FourthHomework28.09.21/ModulesAndFiles/WorkWithArrays.c
@@ -1,0 +1,56 @@
+#include "WorkWithArrays.h"
+#include <stdio.h>
+#include <stdlib.h>
+
+void fillArrayWithRandomElements(int theArray[], int lengthOfTheArray, int mod)
+{
+    for (int i = 0; i < lengthOfTheArray; ++i)
+    {
+        theArray[i] = rand() % mod;
+    }
+}
+
+void printArray(const int theArray[], const int lengthOfTheArray)
+{
+    for (int i = 0; i < lengthOfTheArray; ++i)
+    {
+        printf("%d ", theArray[i]);
+    }
+    printf("\n\n");
+}
+
+bool checkThatArrayIsSorted(const int theArray[], const int lengthOfTheArray)
+{
+    for (int i = 0; i < lengthOfTheArray - 1; ++i)
+    {
+        if (theArray[i] > theArray[i + 1])
+        {
+            return false;
+        }
+    }
+    return true;
+}
+
+bool checkThatAnArraysAreIdentical(const int firstArray[],
+                                   const int secondArray[], const int lengthOfArrays)
+{
+    bool identical = true;
+
+    for (int i = 0; i < lengthOfArrays; ++i)
+    {
+        if (firstArray[i] != secondArray[i])
+        {
+            identical = false;
+            break;
+        }
+    }
+
+    return identical;
+}
+
+void swap(int* firstVariable, int* secondVariable)
+{
+    *firstVariable ^= *secondVariable;
+    *secondVariable ^= *firstVariable;
+    *firstVariable ^= *secondVariable;
+}
