+++ FourthHomework28.09.21/ModulesAndFiles/myQSort.h
@@ -1,3 +1,4 @@
 #pragma once
 
+// The function that sorts the array with the Quick Sort
 void myQSort(int theArray[], const int lengthOfTheArray);
