+++ FourthHomework28.09.21/ModulesAndFiles/testMyQSort.h
@@ -1,3 +1,5 @@
 #pragma once
+#include <stdbool.h>
 
+// Tests for myQSort function
 bool testMyQSort();
